split skiplist source file
