+++ CMakeLists.txt
@@ -52,7 +52,8 @@ set(SOURCES
         engine/kv_engine.cpp
         engine/logger.cpp
         engine/hash_table.cpp
-        engine/skiplist.cpp
+        engine/sorted_collection/skiplist.cpp
+        engine/sorted_collection/rebuilder.cpp
         engine/write_batch.cpp
         engine/dram_allocator.cpp
         engine/pmem_allocator/pmem_allocator.cpp
