+++ engine/kv_engine.cpp
@@ -20,7 +20,8 @@
 #include "configs.hpp"
 #include "dram_allocator.hpp"
 #include "kvdk/engine.hpp"
-#include "skiplist.hpp"
+#include "sorted_collection/iterator.hpp"
+#include "sorted_collection/skiplist.hpp"
 #include "structures.hpp"
 #include "utils/sync_point.hpp"
 #include "utils/utils.hpp"
