+++ engine/kv_engine.hpp
@@ -26,7 +26,8 @@
 #include "logger.hpp"
 #include "pmem_allocator/pmem_allocator.hpp"
 #include "queue.hpp"
-#include "skiplist.hpp"
+#include "sorted_collection/rebuilder.hpp"
+#include "sorted_collection/skiplist.hpp"
 #include "structures.hpp"
 #include "thread_manager.hpp"
 #include "unordered_collection.hpp"
